pyproject.toml
[project]
name = "gcp-testing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain-google-vertexai>=3.2.0",
    "langchain-google-genai>=2.0.5",
    "langgraph>=1.0.5",
    "google-cloud-aiplatform>=1.112.0",
    "python-dotenv>=1.0.0",
    "google-genai>=0.3.0",
]
